+++ IoC.Comparison/TestInfo.cs
@@ -6,7 +6,7 @@
     {
         public TestInfo(
             [NotNull] string name,
-            [NotNull] Action<int, IPerformanceCounter> test)
+            [NotNull] Action<long, IPerformanceCounter> test)
         {
             Name = name ?? throw new ArgumentNullException(nameof(name));
             Test = test ?? throw new ArgumentNullException(nameof(test));
@@ -15,7 +15,7 @@
 
         public string Name { get; }
 
-        public Action<int, IPerformanceCounter> Test { get; }
+        public Action<long, IPerformanceCounter> Test { get; }
 
         public double PerformanceRate { get; set; } = 1;
     }
